+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Harmonize the category name used across the mod.
+
 ## [0.15.1] 2025-07-02
 
 ### Fixed