+++ CHANGELOG.md
@@ -142,7 +142,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   First implementation of the library.
 
-[Unreleased]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.15.1...HEAD
+[Unreleased]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.15.2...HEAD
+[0.15.2]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.15.1...v0.15.2
 [0.15.1]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.13.0...v0.14.0